pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Django project with uv"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2.19,<5.0",
    "djangorestframework>=3.14.0,<4.0",
    "djangorestframework-simplejwt>=5.3.0,<6.0",
    "django-cors-headers>=3.7.0,<4.0",
    "gunicorn>=20.1.0",
    "mysqlclient>=2.2.4,<3.0",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "drf-spectacular>=0.27.0,<1.0",
    "google-cloud-storage>=2.10.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=3.5.0,<4.0",
    "pytest-django>=4.9.0,<5.0",
    "pytest>=8.3.3,<9.0",
]

[tool.uv]
dev-dependencies = [
    "django-debug-toolbar>=3.5.0,<4.0",
    "pytest-django>=4.9.0,<5.0",
    "pytest>=8.3.3,<9.0",
]
